* Merge from SysteMoC--mainline--0.8/SysteMoC--devel--0.9 branch - Fix compiler warning
Patches applied:

 * falk@cs.fau.de--2008-systemoc-intern/SysteMoC--devel--0.9--patch-7
   
 * Merge from SysteMoC--mainline--0.8
 - Fix compiler warning


git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SysteMoC--mainline--0.9--patch-22

+++ pkginclude/systemoc/smoc_multiplex_fifo.hpp
@@ -572,7 +572,7 @@ protected:
 
   /// @brief See smoc_port_in_base_if
   size_t inTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_in_if
   typename this_type::access_type *getReadPortAccess()
@@ -813,7 +813,7 @@ protected:
 
   /// @brief See smoc_port_out_base_if
   size_t outTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_out_if
   AccessImpl *getWritePortAccess()
